+++ Runtime/AsyncImageLoader/FreeImage.cs
@@ -2,7 +2,7 @@ using System;
 using System.Runtime.InteropServices;
 
 public static partial class AsyncImageLoader {
-  public class FreeImage {
+  public static class FreeImage {
     public enum Format {
       FIF_UNKNOWN = -1,
       FIF_BMP = 0,
@@ -46,28 +46,58 @@ public static partial class AsyncImageLoader {
     }
 
     internal enum Type {
+      /// <summary>
+      /// unknown type
+      /// </summary>
       FIT_UNKNOWN = 0,
+      /// <summary>
+      /// standard image : 1-, 4-, 8-, 16-, 24-, 32-bit
+      /// </summary>
       FIT_BITMAP = 1,
+      /// <summary>
+      /// array of unsigned short : unsigned 16-bit
+      /// </summary>
       FIT_UINT16 = 2,
+      /// <summary>
+      /// array of short : signed 16-bit
+      /// </summary>
       FIT_INT16 = 3,
+      /// <summary>
+      /// array of unsigned long : unsigned 32-bit
+      /// </summary>
       FIT_UINT32 = 4,
+      /// <summary>
+      /// array of long : signed 32-bit
+      /// </summary>
       FIT_INT32 = 5,
+      /// <summary>
+      /// array of float : 32-bit IEEE floating point
+      /// </summary>
       FIT_FLOAT = 6,
+      /// <summary>
+      /// array of double : 64-bit IEEE floating point
+      /// </summary>
       FIT_DOUBLE = 7,
+      /// <summary>
+      /// array of FICOMPLEX : 2 x 64-bit IEEE floating point
+      /// </summary>
       FIT_COMPLEX = 8,
+      /// <summary>
+      /// 48-bit RGB image : 3 x 16-bit
+      /// </summary>
       FIT_RGB16 = 9,
+      /// <summary>
+      /// 64-bit RGBA image : 4 x 16-bit
+      /// </summary>
       FIT_RGBA16 = 10,
+      /// <summary>
+      /// 96-bit RGB float image : 3 x 32-bit IEEE floating point
+      /// </summary>
       FIT_RGBF = 11,
+      /// <summary>
+      /// 128-bit RGBA float image : 4 x 32-bit IEEE floating point
+      /// </summary>
       FIT_RGBAF = 12
-    }
-
-    internal enum ColorType {
-      FIC_MINISWHITE = 0,
-      FIC_MINISBLACK = 1,
-      FIC_RGB = 2,
-      FIC_PALETTE = 3,
-      FIC_RGBALPHA = 4,
-      FIC_CMYK = 5
     }
 
     [System.Flags]
@@ -142,6 +172,9 @@ public static partial class AsyncImageLoader {
 
     const string FreeImageLibrary = "FreeImage";
 
+    [DllImport(FreeImageLibrary, EntryPoint = "FreeImage_IsLittleEndian")]
+    internal static extern bool IsLittleEndian();
+
     [DllImport(FreeImageLibrary, EntryPoint = "FreeImage_GetFileTypeFromMemory")]
     internal static extern Format GetFileTypeFromMemory(IntPtr memory, int size);
 
@@ -171,5 +204,14 @@ public static partial class AsyncImageLoader {
 
     [DllImport(FreeImageLibrary, EntryPoint = "FreeImage_GetBPP")]
     internal static extern int GetBPP(IntPtr dib);
+
+    [DllImport(FreeImageLibrary, EntryPoint = "FreeImage_GetBits")]
+    internal static extern IntPtr GetBits(IntPtr dib);
+
+    [DllImport(FreeImageLibrary, EntryPoint = "FreeImage_GetLine")]
+    internal static extern int GetLine(IntPtr dib);
+
+    [DllImport(FreeImageLibrary, EntryPoint = "FreeImage_GetPitch")]
+    internal static extern int GetPitch(IntPtr dib);
   }
 }