+++ Runtime/ImageImporter.cs
@@ -1,9 +1,7 @@
 ﻿using System;
 using System.Runtime.InteropServices;
 using System.Threading.Tasks;
-using Unity.Burst;
 using Unity.Collections;
-using Unity.Collections.LowLevel.Unsafe;
 using Unity.Mathematics;
 using Unity.Jobs;
 using Unity.Profiling;
@@ -11,9 +9,10 @@ using UnityEngine;
 using static Unity.Mathematics.math;
 
 public static partial class AsyncImageLoader {
+  static readonly int MaxTextureSize = SystemInfo.maxTextureSize;
+  static readonly int MaxMipmapLevel = (int)Mathf.Log(SystemInfo.maxTextureSize, 2f) + 1;
+
   class ImageImporter : IDisposable {
-    static readonly int MaxTextureSize = SystemInfo.maxTextureSize;
-    static readonly int MaxMipmapLevel = (int)Mathf.Log(SystemInfo.maxTextureSize, 2f) + 1;
     static readonly ProfilerMarker ConstructorMarker = new ProfilerMarker("ImageImporter.Constructor");
     static readonly ProfilerMarker CreateNewTextureMarker = new ProfilerMarker("ImageImporter.CreateNewTexture");
     static readonly ProfilerMarker LoadIntoTextureMarker = new ProfilerMarker("ImageImporter.LoadIntoTexture");
@@ -21,11 +20,12 @@ public static partial class AsyncImageLoader {
     LoaderSettings _loaderSettings;
 
     IntPtr _bitmap;
-    bool _isBitmapType;
     int _width;
     int _height;
     TextureFormat _textureFormat;
     int _bytesPerPixel;
+    int _imageBitsPerPixel;
+    FreeImage.Type _imageType;
 
     public ImageImporter(byte[] imageData, LoaderSettings loaderSettings) {
       using (ConstructorMarker.Auto()) {
@@ -51,6 +51,8 @@ public static partial class AsyncImageLoader {
           _bitmap = FreeImage.LoadFromMemory(_loaderSettings.format, memoryStream, (int)FreeImage.LoadFlags.JPEG_ACCURATE);
           _width = (int)FreeImage.GetWidth(_bitmap);
           _height = (int)FreeImage.GetHeight(_bitmap);
+          _imageBitsPerPixel = FreeImage.GetBPP(_bitmap);
+          _imageType = FreeImage.GetImageType(_bitmap);
 
           if (_width > MaxTextureSize || _height > MaxTextureSize) {
             Dispose();
@@ -77,19 +79,19 @@ public static partial class AsyncImageLoader {
     }
 
     public void ReinitializeTexture(Texture2D texture) {
+      if ((texture.width, texture.height, texture.format, texture.mipmapCount > 1) != (_width, _height, _textureFormat, _loaderSettings.generateMipmap)) {
 #if UNITY_2021_1_OR_NEWER
-      texture.Reinitialize(_width, _height, _textureFormat, _loaderSettings.generateMipmap);
+        texture.Reinitialize(_width, _height, _textureFormat, _loaderSettings.generateMipmap);
 #else
-      texture.Resize(_width, _height, _textureFormat, _loaderSettings.generateMipmap);
+        texture.Resize(_width, _height, _textureFormat, _loaderSettings.generateMipmap);
 #endif
+      }
     }
 
     public Texture2D CreateNewTexture() {
       using (CreateNewTextureMarker.Auto()) {
         var texture = CreateTexture();
-        var dependency = LoadRawTextureData(texture);
-        if (_isBitmapType) dependency = SwapColorChannel(texture, dependency);
-        dependency = ScheduleMipmapJobForEachLevels(texture, dependency);
+        var dependency = ProcessImage(texture);
         dependency.Complete();
         texture.Apply(false, _loaderSettings.markNonReadable);
         return texture;
@@ -98,9 +100,7 @@ public static partial class AsyncImageLoader {
 
     public async Task<Texture2D> CreateNewTextureAsync() {
       var texture = CreateTexture();
-      var dependency = LoadRawTextureData(texture);
-      if (_isBitmapType) dependency = SwapColorChannel(texture, dependency);
-      dependency = ScheduleMipmapJobForEachLevels(texture, dependency);
+      var dependency = ProcessImage(texture);
       while (!dependency.IsCompleted) await Task.Yield();
       dependency.Complete();
       texture.Apply(false, _loaderSettings.markNonReadable);
@@ -110,9 +110,7 @@ public static partial class AsyncImageLoader {
     public void LoadIntoTexture(Texture2D texture) {
       using (LoadIntoTextureMarker.Auto()) {
         ReinitializeTexture(texture);
-        var dependency = LoadRawTextureData(texture);
-        if (_isBitmapType) dependency = SwapColorChannel(texture, dependency);
-        dependency = ScheduleMipmapJobForEachLevels(texture, dependency);
+        var dependency = ProcessImage(texture);
         dependency.Complete();
         texture.Apply(false, _loaderSettings.markNonReadable);
       }
@@ -120,9 +118,7 @@ public static partial class AsyncImageLoader {
 
     public async Task LoadIntoTextureAsync(Texture2D texture) {
       ReinitializeTexture(texture);
-      var dependency = LoadRawTextureData(texture);
-      if (_isBitmapType) dependency = SwapColorChannel(texture, dependency);
-      dependency = ScheduleMipmapJobForEachLevels(texture, dependency);
+      var dependency = ProcessImage(texture);
       while (!dependency.IsCompleted) await Task.Yield();
       dependency.Complete();
       texture.Apply(false, _loaderSettings.markNonReadable);
@@ -155,15 +151,9 @@ public static partial class AsyncImageLoader {
     }
 
     void DetermineTextureFormat() {
-      var type = FreeImage.GetImageType(_bitmap);
-
-      switch (type) {
+      switch (_imageType) {
         case FreeImage.Type.FIT_BITMAP:
-          var bpp = FreeImage.GetBPP(_bitmap);
-
-          _isBitmapType = true;
-
-          switch (bpp) {
+          switch (_imageBitsPerPixel) {
             case 24:
               _textureFormat = TextureFormat.RGB24;
               _bytesPerPixel = 3;
@@ -173,36 +163,93 @@ public static partial class AsyncImageLoader {
               _bytesPerPixel = 4;
               break;
             default:
-              throw new Exception($"Bitmap bitdepth not supported: {bpp}");
+              throw new Exception($"Bitmap bitdepth not supported: {_imageBitsPerPixel}");
           }
+          break;
+        case FreeImage.Type.FIT_INT16:
+        case FreeImage.Type.FIT_UINT16:
+          _textureFormat = TextureFormat.R16;
+          _bytesPerPixel = 2;
+          break;
+        case FreeImage.Type.FIT_FLOAT:
+          _textureFormat = TextureFormat.RFloat;
+          _bytesPerPixel = 4;
+          break;
+        case FreeImage.Type.FIT_RGB16:
+          _textureFormat = TextureFormat.RGB48;
+          _bytesPerPixel = 6;
+          break;
+        case FreeImage.Type.FIT_RGBA16:
+          _textureFormat = TextureFormat.RGBA64;
+          _bytesPerPixel = 8;
+          break;
+        case FreeImage.Type.FIT_RGBF:
+        case FreeImage.Type.FIT_RGBAF:
+          _textureFormat = TextureFormat.RGBAFloat;
+          _bytesPerPixel = 16;
           break;
         default:
           throw new Exception(
-            $"Image type not supported: {type}. Please submit a new GitHub issue if you want this format to be supported."
+            $"Image type not supported: {_imageType}. Please submit a new GitHub issue if you want this type to be supported."
           );
       }
     }
 
-    public JobHandle LoadRawTextureData(Texture2D texture, JobHandle dependency = default) {
-      var textureData = GetTextureData(texture, 0);
-      var loadRawTextureDataJob = new LoadRawTextureDataJob {
-        width = _width,
-        bytesPerPixel = _bytesPerPixel,
-        bitmap = _bitmap,
-        textureData = textureData,
-      };
-
-      return loadRawTextureDataJob.Schedule(dependency);
+    public JobHandle ProcessImage(Texture2D texture) {
+      JobHandle dependency = default;
+      dependency = ScheduleLoadTextureDataJob(texture, dependency);
+      dependency = ScheduleMipmapJobForEachLevels(texture, dependency);
+      return dependency;
     }
 
-    public JobHandle SwapColorChannel(Texture2D texture, JobHandle dependency = default) {
+    public unsafe JobHandle ScheduleLoadTextureDataJob(Texture2D texture, JobHandle dependency = default) {
       var textureData = GetTextureData(texture, 0);
-      var swapColorChannelJob = new SwapColorChannelJob {
-        channelCount = _textureFormat == TextureFormat.RGBA32 ? RGBA32Pixel.ChannelCount() : RGB24Pixel.ChannelCount(),
-        textureData = textureData,
-      };
+      var bitsPtr = FreeImage.GetBits(_bitmap);
+      var bytesPerLine = FreeImage.GetLine(_bitmap);
+      var bytesPerScanline = FreeImage.GetPitch(_bitmap);
 
-      return swapColorChannelJob.Schedule(_width * _height, min(_width, _height), dependency);
+      if (FreeImage.IsLittleEndian() && _imageType == FreeImage.Type.FIT_BITMAP) {
+        switch (_imageBitsPerPixel) {
+          case 24:
+            var transferBGR24ImageToRGB24TextureJob = new TransferBGR24ImageToRGB24TextureJob {
+              bytesPerScanline = bytesPerScanline,
+              width = _width,
+              bitsPtr = bitsPtr,
+              textureData = textureData,
+            };
+
+            return transferBGR24ImageToRGB24TextureJob.Schedule(_height, 1, dependency);
+          case 32:
+            var transferBGRA32ImageToRGBA32TextureJob = new TransferBGRA32ImageToRGBA32TextureJob {
+              bytesPerScanline = bytesPerScanline,
+              width = _width,
+              bitsPtr = bitsPtr,
+              textureData = textureData,
+            };
+
+            return transferBGRA32ImageToRGBA32TextureJob.Schedule(_height, 1, dependency);
+          default:
+            throw new Exception($"Bitmap bitdepth not supported: {_imageBitsPerPixel}");
+        }
+      } else if (_imageType == FreeImage.Type.FIT_RGBF) {
+        var transferRGBFloatImageToRGBAFloatTextureJob = new TransferRGBFloatImageToRGBAFloatTextureJob {
+          bytesPerScanline = bytesPerScanline,
+          width = _width,
+          bitsPtr = bitsPtr,
+          textureData = textureData,
+        };
+
+        return transferRGBFloatImageToRGBAFloatTextureJob.Schedule(_height, 1, dependency);
+      } else {
+        var transferImageToTextureJob = new TransferImageToTextureJob {
+          bytesPerLine = bytesPerLine,
+          bytesPerScanline = bytesPerScanline,
+          bitsPtr = bitsPtr,
+          textureData = textureData,
+        };
+
+        return transferImageToTextureJob.Schedule(_height, 1, dependency);
+      }
     }
 
     public JobHandle ScheduleMipmapJobForEachLevels(Texture2D texture, JobHandle dependency = default) {
@@ -213,26 +260,44 @@ public static partial class AsyncImageLoader {
         var nextMipmapSize = GetMipmapSize(mipmapLevel);
         var nextMipmapSlice = GetTextureData(texture, mipmapLevel);
 
-        switch (_textureFormat) {
-          case TextureFormat.RGBA32:
-            dependency = ScheduleMipmapJobForEachChannels<byte, RGBA32Pixel, ByteChannelMipmapJob>(
-              RGBA32Pixel.ChannelCount(), RGBA32Pixel.ChannelByteCount(),
-              mipmapSlice, mipmapSize,
-              nextMipmapSlice, nextMipmapSize,
-              dependency
-            );
-            break;
-          case TextureFormat.RGB24:
-            dependency = ScheduleMipmapJobForEachChannels<byte, RGB24Pixel, ByteChannelMipmapJob>(
+        dependency = _textureFormat switch {
+          TextureFormat.R16 => ScheduleMipmapJobForEachChannels<short, R16Pixel, ShortChannelMipmapJob>(
+            R16Pixel.ChannelCount(), R16Pixel.ChannelByteCount(),
+            mipmapSlice, mipmapSize, nextMipmapSlice, nextMipmapSize,
+            dependency
+          ),
+          TextureFormat.RFloat => ScheduleMipmapJobForEachChannels<float, RFloatPixel, FloatChannelMipmapJob>(
+            RFloatPixel.ChannelCount(), RFloatPixel.ChannelByteCount(),
+            mipmapSlice, mipmapSize, nextMipmapSlice, nextMipmapSize,
+            dependency
+          ),
+          TextureFormat.RGB24 => ScheduleMipmapJobForEachChannels<byte, RGB24Pixel, ByteChannelMipmapJob>(
               RGB24Pixel.ChannelCount(), RGB24Pixel.ChannelByteCount(),
-              mipmapSlice, mipmapSize,
-              nextMipmapSlice, nextMipmapSize,
+              mipmapSlice, mipmapSize, nextMipmapSlice, nextMipmapSize,
               dependency
-            );
-            break;
-          default:
-            throw new NotSupportedException($"Texture format not supported: {_textureFormat}");
-        }
+            ),
+          TextureFormat.RGB48 => ScheduleMipmapJobForEachChannels<short, RGB48Pixel, ShortChannelMipmapJob>(
+            RGB48Pixel.ChannelCount(), RGB48Pixel.ChannelByteCount(),
+            mipmapSlice, mipmapSize, nextMipmapSlice, nextMipmapSize,
+            dependency
+          ),
+          TextureFormat.RGBA32 => ScheduleMipmapJobForEachChannels<byte, RGBA32Pixel, ByteChannelMipmapJob>(
+            RGBA32Pixel.ChannelCount(), RGBA32Pixel.ChannelByteCount(),
+            mipmapSlice, mipmapSize, nextMipmapSlice, nextMipmapSize,
+            dependency
+          ),
+          TextureFormat.RGBA64 => ScheduleMipmapJobForEachChannels<short, RGBA64Pixel, ShortChannelMipmapJob>(
+            RGBA64Pixel.ChannelCount(), RGBA64Pixel.ChannelByteCount(),
+            mipmapSlice, mipmapSize, nextMipmapSlice, nextMipmapSize,
+            dependency
+          ),
+          TextureFormat.RGBAFloat => ScheduleMipmapJobForEachChannels<float, RGBAFloatPixel, FloatChannelMipmapJob>(
+            RGBAFloatPixel.ChannelCount(), RGBAFloatPixel.ChannelByteCount(),
+            mipmapSlice, mipmapSize, nextMipmapSlice, nextMipmapSize,
+            dependency
+          ),
+          _ => throw new NotSupportedException($"Texture format not supported: {_textureFormat}"),
+        };
 
         mipmapSize = nextMipmapSize;
         mipmapSlice = nextMipmapSlice;
@@ -286,39 +351,6 @@ public static partial class AsyncImageLoader {
 
       return new NativeSlice<byte>(rawTextureData, mipmapOffset, mipmapByteSize);
 #pragma warning restore CS0162
-    }
-
-    struct LoadRawTextureDataJob : IJob {
-      public int width;
-      public int bytesPerPixel;
-
-      [NativeDisableUnsafePtrRestriction]
-      public IntPtr bitmap;
-
-      [WriteOnly] public NativeSlice<byte> textureData;
-
-      public unsafe void Execute() {
-        FreeImage.ConvertToRawBits(
-          (IntPtr)textureData.GetUnsafePtr(), bitmap,
-          bytesPerPixel * width,
-          (uint)(bytesPerPixel * 8),
-          0, 0, 0, false
-        );
-      }
-    }
-
-    [BurstCompile(CompileSynchronously = true)]
-    struct SwapColorChannelJob : IJobParallelFor {
-      public int channelCount;
-
-      [NativeDisableParallelForRestriction]
-      public NativeSlice<byte> textureData;
-
-      public void Execute(int index) {
-        var temp = textureData[mad(channelCount, index, 0)];
-        textureData[mad(channelCount, index, 0)] = textureData[mad(channelCount, index, 2)];
-        textureData[mad(channelCount, index, 2)] = temp;
-      }
     }
   }
 }