+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## [Unreleased]
+### Enhancement
+- Add support for non-bitmap image type.
+
 ### Fixed
-- Fix accuracy loss when generating mipmap (#7)
+- Fix accuracy loss when generating mipmap (#7).
 
 ## [0.1.2] - 2022-02-17
 ### Fixed